+++ packages/abap-deploy-config-inquirer/test/service-provider-utils/transport-config.test.ts
@@ -14,7 +14,6 @@ describe('getTransportConfigInstance', () => {
     it('should return the dummy instance of TransportConfig', async () => {
         const transportConfigResult = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: true,
             credentials: {}
         });
         expect(transportConfigResult.transportConfig?.getPackage()).toBe(undefined);
@@ -41,7 +40,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
 
@@ -69,7 +67,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
 
@@ -87,7 +84,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult2 = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
 
@@ -112,7 +108,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
 
@@ -137,7 +132,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
         expect(transportConfigResult.transportConfigNeedsCreds).toBe(true);
@@ -156,7 +150,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResultWithoutHeaders = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
         expect(transportConfigResultWithoutHeaders.transportConfigNeedsCreds).toBe(false);
@@ -175,7 +168,6 @@ describe('getTransportConfigInstance', () => {
 
         const transportConfigResult2 = await getTransportConfigInstance({
             backendTarget: undefined,
-            scp: false,
             credentials: {}
         });
         expect(transportConfigResult2.transportConfigNeedsCreds).toBe(false);