+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v4/registry.ts
@@ -17,6 +17,7 @@ import { ToggleSemanticDateRangeFilterBar } from './lr-enable-semantic-date-rang
 import { EnableTableEmptyRowModeQuickAction } from './op-enable-empty-row-mode';
 import { AddNewAnnotationFile } from '../common/add-new-annotation-file';
 import { EnableVariantManagementQuickAction } from './enable-variant-management';
+import { AddNewSubpage } from '../fe-v4/add-new-subpage';
 
 type PageName = 'listReport' | 'objectPage';
 
@@ -50,7 +51,8 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction,
                         EnableTableFilteringQuickAction,
-                        AddNewAnnotationFile
+                        AddNewAnnotationFile,
+                        AddNewSubpage
                     ],
                     view,
                     key: name + index
@@ -68,7 +70,8 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction,
                         EnableTableEmptyRowModeQuickAction,
-                        AddNewAnnotationFile
+                        AddNewAnnotationFile,
+                        AddNewSubpage
                     ],
                     view,
                     key: name + index