+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.0.10
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.10
+
+## 0.0.9
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.9
+
 ## 0.0.8
 
 ### Patch Changes